Changed eval-expr to return Integer structs.
This comes with a lot of refactoring of the logic for evaling exprs, as the binary ops in the op table need to be able to pull apart the Integer values to get the actual numbers within, then place the result into a new Integer value. Also note that the unit tests are broken (for now) due to the refactor.